@@ -2,38 +2,46 @@
   "name": "yahtzee-online",
   "version": "1.0.0",
   "description": "Full Stack React App with OAuth React and Typescript",
-  "main": "server/server.js",
+  "main": "./server/server.ts",
   "scripts": {
     "test": "echo \"Error: no tests installed\" && exit 1",
-    "prestart": "npm run build",
-    "build": "tsc",
-    "start": "if-env NODE_ENV=production && npm run start:prod || npm run start:dev",
-    "start:prod": "node build-ts server/server.ts",
+    "prestart": "npm run tsc",
+    "start": "if-env NODE_ENV=production && npm run start:prod || npm run start:dev && npm run dev",
+    "start:prod": "npm run tsc && NODE_ENV=production node ./server/dist/server.js",
     "start:dev": "concurrently \"nodemon --ignore 'client/*'\" \"npm run client\"",
+    "client": "cd client && npm run start",
     "seed": "node server/scripts/seedDB.js",
     "install": "cd client && npm install",
     "build-ts": "cd client && npm run build",
     "heroku-postbuild": "npm run build",
-    "tsc": "tsc",
-    "dev": "ts-node-dev --respawn --transpileOnly ./src/index.ts"
+    "tsc": "tsc --outdir ./server/dist",
+    "dev": "ts-node-dev --respawn --watch server/**/*.ts --transpileOnly ./server/server.ts"
   },
   "author": "Jon R Jackson",
   "license": "ISC",
   "devDependencies": {
-    "nodemon": "^1.19.4"
+    "nodemon": "^1.19.4",
+    "ts-node-dev": "^1.0.0-pre.44"
   },
   "dependencies": {
-    "axios": "^0.19.0",
+    "@types/axios": "^0.14.0",
+    "@types/express": "^4.17.2",
+    "@types/express-session": "^1.15.15",
+    "@types/jsonwebtoken": "^8.3.5",
+    "@types/mongoose": "^5.5.29",
+    "@types/node": "^12.12.5",
     "bcrypt": "^3.0.2",
     "bcryptjs": "^2.4.3",
-    "connect-mongo": "^3.0.0",
+    "concurrently": "^5.0.0",
+    "connect-mongo": "^3.1.2",
     "dotenv": "^8.2.0",
     "express": "^4.16.4",
     "express-session": "^1.15.6",
     "if-env": "^1.0.4",
     "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.3.4",
+    "mongoose": "^5.7.8",
     "morgan": "^1.9.1",
-    "typescript": "^3.6.4"
+    "ts-node": "^8.4.1",
+    "typescript": "^3.7.2"
   }
 }
