@@ -1,37 +1,36 @@
-const db = require("../models");
+const db = require('../models');
+
+import { IPlayer } from '../models/player';
+import { Request, Response } from 'express';
+import { Error } from 'mongoose';
 
 // Defining methods for the ModelsController
 module.exports = {
-  findAll: function(req, res) {
-    db.Model
-      .find(req.query)
+  findAll: function(req: Request, res: Response) {
+    db.Player.find(req.query)
       .sort({ date: -1 })
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
+      .then((dbPlayer: IPlayer) => res.json(dbPlayer))
+      .catch((err: Error) => res.status(422).json(err));
   },
-  findById: function(req, res) {
-    db.Model
-      .findById(req.params.id)
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
+  findById: function(req: Request, res: Response) {
+    db.Player.findById(req.params.id)
+      .then((dbPlayer: IPlayer) => res.json(dbPlayer))
+      .catch((err: Error) => res.status(422).json(err));
   },
-  create: function(req, res) {
-    db.Model
-      .create(req.body)
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
+  create: function(req: Request, res: Response) {
+    db.Player.create(req.body)
+      .then((dbPlayer: IPlayer) => res.json(dbPlayer))
+      .catch((err: Error) => res.status(422).json(err));
   },
-  update: function(req, res) {
-    db.Model
-      .findOneAndUpdate({ _id: req.params.id }, req.body)
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
+  update: function(req: Request, res: Response) {
+    db.Player.findOneAndUpdate({ _id: req.params.id }, req.body)
+      .then((dbPlayer: IPlayer) => res.json(dbPlayer))
+      .catch((err: Error) => res.status(422).json(err));
   },
-  remove: function(req, res) {
-    db.Model
-      .findById({ _id: req.params.id })
-      .then(dbModel => dbModel.remove())
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
-  }
+  remove: function(req: Request, res: Response) {
+    db.Player.findById({ _id: req.params.id })
+      // .then((dbPlayer: IPlayer) => dbPlayer.remove())
+      .then((dbPlayer: IPlayer) => res.json(dbPlayer))
+      .catch((err: Error) => res.status(422).json(err));
+  },
 };
