@@ -1,10 +1,12 @@
-const mongoose = require("mongoose");
+import { MongooseDocument } from 'mongoose';
+
+const mongoose = require('mongoose');
 const Schema = mongoose.Schema;
 
-const playerSchema = new Schema({
+export interface IPlayer extends MongooseDocument {}
 
-});
+const playerSchema = new Schema({});
 
-const Player = mongoose.model("Player", playerSchema);
+const Player = mongoose.model('Player', playerSchema);
 
 module.exports = Player;
