@@ -1,12 +1,13 @@
-import React, { ReactPropTypes, FC } from "react";
+import React, { FC, ReactElement } from 'react';
 
-interface CardProps extends ReactPropTypes {
+interface CardProps {
   title: string;
+  children: ReactElement[] | ReactElement;
 }
 
 export const Card: FC<CardProps> = props => (
   <div className="card">
-    <div className="card-header bg-primary" style={{ color: "#fff" }}>
+    <div className="card-header bg-primary" style={{ color: '#fff' }}>
       <h5>{props.title}</h5>
     </div>
     <div className="card-body">{props.children}</div>
