@@ -1,12 +1,11 @@
-import React, { ReactPropTypes, FC } from "react";
+import React, { FC, ReactNode, ReactNodeArray } from 'react';
 
-interface JumbotronProps extends ReactPropTypes {}
+interface JumbotronProps {
+  children: ReactNode | ReactNodeArray;
+}
 
 const Jumbotron: FC<JumbotronProps> = ({ children }) => (
-  <div
-    style={{ height: 300, clear: "both", paddingTop: 120, textAlign: "center" }}
-    className="jumbotron"
-  >
+  <div style={{ height: 300, clear: 'both', paddingTop: 120, textAlign: 'center' }} className="jumbotron">
     {children}
   </div>
 );
