@@ -1,11 +1,11 @@
-import React, { Component, ChangeEvent } from "react";
-import { Redirect, Link } from "react-router-dom";
-import { Container, Row, Col } from "../../components/Grid";
-import { Card } from "../../components/Card";
-import { Input, FormBtn } from "../../components/Form";
-import AUTH from "../../utils/AUTH";
+import React, { Component, ChangeEvent } from 'react';
+import { Redirect, Link } from 'react-router-dom';
+import { Container } from '../../components/Grid';
+import { Card } from '../../components/Card';
+import { Input, FormBtn } from '../../components/Form';
+import AUTH from '../../utils/AUTH';
 
-import "./Auth.css";
+import './Auth.css';
 
 interface Props {}
 interface State {
@@ -22,12 +22,12 @@ class SignupForm extends Component<Props, State> {
     super(props);
 
     this.state = {
-      firstName: "",
-      lastName: "",
-      username: "",
-      password: "",
-      confirmPassword: "",
-      redirectTo: null
+      firstName: '',
+      lastName: '',
+      username: '',
+      password: '',
+      confirmPassword: '',
+      redirectTo: null,
     };
   }
 
@@ -35,27 +35,27 @@ class SignupForm extends Component<Props, State> {
     const { name } = event.target;
     this.setState((prevState: State) => ({
       ...prevState,
-      [event.target.name]: event.target.value
+      [event.target.name]: event.target.value,
     }));
   };
 
-  handleSubmit = event => {
+  handleSubmit = (event: MouseEvent) => {
     event.preventDefault();
     // TODO - validate!
     AUTH.signup({
       firstName: this.state.firstName,
       lastName: this.state.lastName,
       username: this.state.username,
-      password: this.state.password
+      password: this.state.password,
     }).then(response => {
       console.log(response);
       if (!response.data.error) {
-        console.log("youre good");
+        console.log('youre good');
         this.setState({
-          redirectTo: "/"
+          redirectTo: '/',
         });
       } else {
-        console.log("duplicate");
+        console.log('duplicate');
       }
     });
   };
@@ -70,33 +70,13 @@ class SignupForm extends Component<Props, State> {
         <Card title="Lets Play Yahtzee!">
           <form style={{ marginTop: 10 }}>
             <label htmlFor="username">First name: </label>
-            <Input
-              type="text"
-              name="firstName"
-              value={this.state.firstName}
-              onChange={this.handleChange}
-            />
+            <Input type="text" name="firstName" value={this.state.firstName} onChange={this.handleChange} />
             <label htmlFor="username">Last name: </label>
-            <Input
-              type="text"
-              name="lastName"
-              value={this.state.lastName}
-              onChange={this.handleChange}
-            />
+            <Input type="text" name="lastName" value={this.state.lastName} onChange={this.handleChange} />
             <label htmlFor="username">Username: </label>
-            <Input
-              type="text"
-              name="username"
-              value={this.state.username}
-              onChange={this.handleChange}
-            />
+            <Input type="text" name="username" value={this.state.username} onChange={this.handleChange} />
             <label htmlFor="password">Password: </label>
-            <Input
-              type="password"
-              name="password"
-              value={this.state.password}
-              onChange={this.handleChange}
-            />
+            <Input type="password" name="password" value={this.state.password} onChange={this.handleChange} />
             <label htmlFor="confirmPassword">Confirm Password: </label>
             <Input
               type="password"
