@@ -1,1 +1,1 @@
-export { default } from "./NoMatch.js";
+export { default } from './NoMatch';
