@@ -1,22 +1,21 @@
-import React, { Component } from "react";
-import { Link } from "react-router-dom";
+import React, { Component } from 'react';
+// import { Link } from "react-router-dom";
 
-import API from "../../utils/API";
-
+import API from '../../utils/API';
+import { AxiosResponse } from 'axios';
 
 class Yahtzee extends Component {
   state = {
     players: [],
-    socket: {}
+    socket: {},
   };
 
   componentDidMount() {
+    API.checkConnection().then((res: AxiosResponse) => console.log(res));
   }
 
   render() {
-    return (
-      <div></div>
-    );
+    return <div></div>;
   }
 }
 
