@@ -1,5 +1,7 @@
-import axios from "axios";
+import axios from 'axios';
 
 export default {
-  
+  checkConnection: () => {
+    return axios.get('/api');
+  },
 };
