@@ -1,4 +1,4 @@
 module.exports = {
-  Player: require("./player"),
-  User: require("./user")
+  Player: require('./player'),
+  User: require('./user'),
 };
